Fixed more compiler errors

+++ RailRoad.java
@@ -20,7 +20,7 @@ class RailRoad extends Property
       int toll;
       int rails;
       // rails = number of railroads owned by this.owner
-      toll = 25 * Math.pow(2, rails - 1);
+      toll = (int) (25 * Math.pow(2, rails - 1));
       p.payment(-toll);
       owner.payment(toll);
    }
